pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PhotoChart"
version = "0.0.0"
description = "A tool for managing and organizing photo collections."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "" }
]
dependencies = [
  "fuzzywuzzy",
  "pandas",
  "tabulate",
  "python-Levenshtein",
  "openpyxl",
  "odfpy",
  "rich",
  "plotly",
  "dash",
  "dash-bootstrap-components",
  "kaleido",
  "xlrd>=2.0.1",
  "Django>=4.2.0",
  "djangorestframework>=3.14.0",
  "django-filter>=23.0",
  "django-cors-headers>=4.0.0",
  "dj-database-url>=2.0.0",
  "flatten_dict==0.4.2",
  "matplotlib>=3.9",
  "matplotlib-inline==0.1.7",
  "PyYAML==6.0.1",
  "setuptools==70.1.0",
  "sysrsync==1.1.1",
  "tqdm==4.66.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "pytest-mock>=3.10.0",
  "pytest-asyncio>=0.21.0",
  "freezegun>=1.2.0",
  "pre-commit>=4.5.1",
  "black==25.*",
]
docs = [
  "sphinx==7.2.6",
  "sphinx-rtd-theme==2.0.0",
  "sphinx-autodoc-typehints==1.25.2",
]
nef = [
  "rawpy>=0.20.0",
  "Pillow>=10.0.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["cli*", "backend*", "photochart"]
exclude = ["scripts*"]

[project.scripts]
pchart = "cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=backend",
    "--cov=cli",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
